特效部分封装了一层;支持了技能同时索敌己方和敌方

+++ Assets/Scripts/Core/Unit/Unit.cs
@@ -65,7 +65,7 @@ public class Unit
 
     public float PowerSpeed;
 
-    public float Team;
+    public int Team;
 
     public int Weight;
     public int WeightBase, WeightAdd;
